+++ PhpSerializerNET.Test/Serialize/IPhpObjectSerialization.cs
@@ -4,21 +4,19 @@
   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 **/
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using PhpSerializerNET.Test.DataTypes;
 
 namespace PhpSerializerNET.Test {
 	[TestClass]
 	public class IPhpObjectSerializationTest {
-		public class MyPhpObject : IPhpObject {
-			public string GetClassName() => "MyPhpObject";
-			public void SetClassName(string className) {}
-			public string Foo {get;set;}
-		}
 
 		[TestMethod]
 		public void SerializeIPhpObject() {
+			var data = new MyPhpObject() { Foo = "" };
+			data.SetClassName("MyPhpObject");
 			Assert.AreEqual( // strings:
 				"O:11:\"MyPhpObject\":1:{s:3:\"Foo\";s:0:\"\";}",
-				PhpSerialization.Serialize( new MyPhpObject() {Foo =""})
+				PhpSerialization.Serialize(data)
 			);
 		}
 	}